Use startActivityForResult() when launching OpenVPN for Android
The previous use of startActivity() didn't allow OpenVPN for Android to
see which application attempted to start a profile, causing the
whitelist to not work properly. By using startActivityForResult(),
OpenVPN for Android is able to find the caller via getCallingPackage()
and whitelist Fwknop2 so that it can in the future automatically connect
to a VPN profile after sending the SPA.

Tested on LineageOS 15.1 (Android 8.1).

Signed-off-by: Matt Merhar <mattmerhar@protonmail.com>